requirements.txt
Django>=2.2.4,<2.3.0
djangorestframework>3.9.2.2,<3.10.0
django-filter==2.2.0
psycopg2==2.8.3